Rename modules to include leading underscores for consistency and update imports
