+++ braintools/visualize/_animation.py
